+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Enhancements
+
+— Add translate full sentence by click
 — Load subtitles for Onvix films
 — Add extension icon